Use one import for axios in actions
