+++ src/actions/signupAction.js
@@ -1,6 +1,3 @@
-import axios from 'axios';
-import { toast } from 'react-toastify';
-
 const REGISTER_USER = 'REGISTER_USER';
 const REGISTER_USER_SUCCESS = 'REGISTER_USER_SUCCESS';
 const REGISTER_USER_FAILURE = 'REGISTER_USER_FAILURE';
@@ -19,39 +16,3 @@ export const registerUserfailure = error => ({
   type: REGISTER_USER_FAILURE,
   payload: { error: error.message },
 });
-
-export const signupAction = newUser => (dispatch) => {
-  dispatch(registerUser());
-  return axios
-    .post('https://capstone-api-v1.herokuapp.com/signup', {user: newUser})
-    .then((response) => {
-      toast.dismiss();
-      dispatch(registerUserSucess(response));
-      if (response.status === 201) {
-        toast.success('Account created successfully', {
-          autoClose: 12000,
-          hideProgressBar: false,
-        });
-      } else if (response.status === 200) {
-        toast.error(`${response.message}`, {
-          autoClose: 3000,
-          hideProgressBar: true,
-        });
-      }
-    })
-    .catch(
-      (error) => {
-        dispatch({
-          type: REGISTER_USER_FAILURE,
-          payload: error.response,
-        });
-        toast.dismiss();
-        toast.error(`${error.response}`, {
-          autoClose: 3000,
-          hideProgressBar: true,
-        });
-      },
-    );
-};
-
-export default signupAction;
